Install the `compile_betterproto` command more cleanly
Instead of hijacking the `build` class by importing it in our command
we use a different entry point,
`setuptools.finalize_distribution_options`, to add our build
sub-command.

This is still not a 100% clean, as that entry point is not specifically
for this purpose, but it is still cleaner than the previous approach.

For more details, see:

* https://github.com/pypa/setuptools/discussions/3180#discussioncomment-2422328
* https://setuptools.pypa.io/en/latest/userguide/extension.html#customizing-distribution-options
* https://github.com/abravalheri/experiment-setuptools-plugin

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>