Properly include generated Python files in the wheel
When building the wheel, the setuptools file discovery feature runs
before the `compile_betterproto` sub-command is run, so the files are
not included in the resulting wheel.

To overcome this issue we use a hack, by running the compile command
early when the distribution options are being finalized, we make sure
the files are present by the time the file discovery runs.

This is very hacky though, and even when it seems to work well in all
tested cases (calling setuptools` `setup()` directly, `build`,
`pip install`, `pip install -e`), I wouldn't be very surprised if it
breaks in the future, so we probably will have to have an eye on it.

Unfortunately it seems there is no easy way to do this with setuptools:

* https://github.com/pypa/setuptools/discussions/3180#discussioncomment-2422328
* https://github.com/pypa/setuptools/issues/2591

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>